feat(components): stats modal
